requirements.txt
pandas
openpyxl
streamlit
python-dateutil
PyYAML
reportlab
yagmail
python-magic
xlsxwriter
